segmind/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
chromadb==0.4.24
supabase==2.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.4